Ensure that Upcoming TBA camps are filtered by region
In order to put TBA camps into the Upcoming list, it's necessary to do a
separate query and tack them onto the template global. Previously, this
additional query was not being filtered by 'region'.

This changeset also fixes an issue in the add-to-groups_template logic, where
the total group count offsets were off by one because of the way that array
indexes start at 0 rather than 1.

See #67