Don't append TBDs when region and date filter are both set to 'all'
When region and date are both set to 'all' (ie, results are "unfiltered"),
we have a special case: camps are sorted by date_created, which includes
TBD camps, so there's no need to append them. However, when the date
filter is set to 'all' and the region filter is present, the sort
order switches to be by the camp date, which means that appending is
necessary.

See #84, #96